Implement ECS: Introduce AwsClient hierarchy and AwsClientConfigurator
